fabric-1.19.3: rename package references
