+++ fabric-1.19.3/src/main/java/org/dynmap/fabric_1_19_3/DynmapPlugin.java
@@ -1,4 +1,4 @@
-package org.dynmap.fabric_1_19_1;
+package org.dynmap.fabric_1_19_3;
 
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
@@ -36,16 +36,16 @@ import org.dynmap.common.DynmapCommandSender;
 import org.dynmap.common.DynmapListenerManager;
 import org.dynmap.common.DynmapPlayer;
 import org.dynmap.common.chunk.GenericChunkCache;
-import org.dynmap.fabric_1_19_1.command.DmapCommand;
-import org.dynmap.fabric_1_19_1.command.DmarkerCommand;
-import org.dynmap.fabric_1_19_1.command.DynmapCommand;
-import org.dynmap.fabric_1_19_1.command.DynmapExpCommand;
-import org.dynmap.fabric_1_19_1.event.BlockEvents;
-import org.dynmap.fabric_1_19_1.event.CustomServerChunkEvents;
-import org.dynmap.fabric_1_19_1.event.CustomServerLifecycleEvents;
-import org.dynmap.fabric_1_19_1.event.PlayerEvents;
-import org.dynmap.fabric_1_19_1.mixin.BiomeEffectsAccessor;
-import org.dynmap.fabric_1_19_1.permissions.*;
+import org.dynmap.fabric_1_19_3.command.DmapCommand;
+import org.dynmap.fabric_1_19_3.command.DmarkerCommand;
+import org.dynmap.fabric_1_19_3.command.DynmapCommand;
+import org.dynmap.fabric_1_19_3.command.DynmapExpCommand;
+import org.dynmap.fabric_1_19_3.event.BlockEvents;
+import org.dynmap.fabric_1_19_3.event.CustomServerChunkEvents;
+import org.dynmap.fabric_1_19_3.event.CustomServerLifecycleEvents;
+import org.dynmap.fabric_1_19_3.event.PlayerEvents;
+import org.dynmap.fabric_1_19_3.mixin.BiomeEffectsAccessor;
+import org.dynmap.fabric_1_19_3.permissions.*;
 import org.dynmap.permissions.PermissionsHandler;
 import org.dynmap.renderer.DynmapBlockState;
