+++ fabric-1.19.3/src/main/java/org/dynmap/fabric_1_19_3/FabricMapChunkCache.java
@@ -1,4 +1,4 @@
-package org.dynmap.fabric_1_19_1;
+package org.dynmap.fabric_1_19_3;
 
 import net.minecraft.nbt.*;
 import net.minecraft.server.world.ServerChunkManager;
