+++ fabric-1.19.3/src/main/java/org/dynmap/fabric_1_19_3/access/ProtoChunkAccessor.java
@@ -1,4 +1,4 @@
-package org.dynmap.fabric_1_19_1.access;
+package org.dynmap.fabric_1_19_3.access;
 
 public interface ProtoChunkAccessor {
     boolean getTouchedByWorldGen();
