+++ fabric-1.19.3/src/main/java/org/dynmap/fabric_1_19_3/mixin/MinecraftServerMixin.java
@@ -1,7 +1,7 @@
-package org.dynmap.fabric_1_19_1.mixin;
+package org.dynmap.fabric_1_19_3.mixin;
 
 import net.minecraft.server.MinecraftServer;
-import org.dynmap.fabric_1_19_1.event.CustomServerLifecycleEvents;
+import org.dynmap.fabric_1_19_3.event.CustomServerLifecycleEvents;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
