+++ fabric-1.19.3/src/main/java/org/dynmap/fabric_1_19_3/mixin/ProtoChunkMixin.java
@@ -1,9 +1,9 @@
-package org.dynmap.fabric_1_19_1.mixin;
+package org.dynmap.fabric_1_19_3.mixin;
 
 import net.minecraft.block.BlockState;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.chunk.ProtoChunk;
-import org.dynmap.fabric_1_19_1.access.ProtoChunkAccessor;
+import org.dynmap.fabric_1_19_3.access.ProtoChunkAccessor;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
