+++ fabric-1.19.3/src/main/java/org/dynmap/fabric_1_19_3/mixin/ServerPlayNetworkHandlerMixin.java
@@ -1,4 +1,4 @@
-package org.dynmap.fabric_1_19_1.mixin;
+package org.dynmap.fabric_1_19_3.mixin;
 
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntity;
@@ -17,8 +17,8 @@ import net.minecraft.util.math.BlockPos;
 import java.util.List;
 
 import net.minecraft.util.registry.RegistryKey;
-import org.dynmap.fabric_1_19_1.event.BlockEvents;
-import org.dynmap.fabric_1_19_1.event.ServerChatEvents;
+import org.dynmap.fabric_1_19_3.event.BlockEvents;
+import org.dynmap.fabric_1_19_3.event.ServerChatEvents;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
