+++ fabric-1.19.3/src/main/java/org/dynmap/fabric_1_19_3/mixin/WorldChunkMixin.java
@@ -1,10 +1,10 @@
-package org.dynmap.fabric_1_19_1.mixin;
+package org.dynmap.fabric_1_19_3.mixin;
 
 import net.minecraft.block.BlockState;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraft.world.chunk.WorldChunk;
-import org.dynmap.fabric_1_19_1.event.BlockEvents;
+import org.dynmap.fabric_1_19_3.event.BlockEvents;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
