+++ fabric-1.19.3/src/main/java/org/dynmap/fabric_1_19_3/permissions/OpPermissions.java
@@ -1,8 +1,8 @@
-package org.dynmap.fabric_1_19_1.permissions;
+package org.dynmap.fabric_1_19_3.permissions;
 
 import net.minecraft.entity.player.PlayerEntity;
 import org.dynmap.Log;
-import org.dynmap.fabric_1_19_1.DynmapPlugin;
+import org.dynmap.fabric_1_19_3.DynmapPlugin;
 
 import java.util.HashSet;
 import java.util.Set;
