+++ fabric-1.19.3/src/main/java/org/dynmap/fabric_1_19_3/permissions/PermissionProvider.java
@@ -1,4 +1,4 @@
-package org.dynmap.fabric_1_19_1.permissions;
+package org.dynmap.fabric_1_19_3.permissions;
 
 import net.minecraft.entity.player.PlayerEntity;
