+++ fabric-1.19.3/src/main/resources/dynmap.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "org.dynmap.fabric_1_19_1.mixin",
+  "package": "org.dynmap.fabric_1_19_3.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
     "BiomeEffectsAccessor",
